usuwanie redundant kopii mapy i world

+++ src/main/java/org/app/MainApp/MainAppController.java
@@ -38,7 +38,7 @@ public class MainAppController {
 
 
     void UpdateSettings() {
-        settings.setAntCircleRadius((int) AntSizeSlide.getValue());
+        settings.setAntCircleRadius(AntSizeSlide.getValue());
     }
 
     @FXML
